Compel_UFFD: Initial code for handling uffd from remote end
This change includes the base code for handling uffd from
remote side. As of this commit, the code is capable of requesting
uffd on the remote side. The remote side will just send back
a dummy page filled with the character 'A'.

Further changes shall be added in the upcoming commits