Synchronous write page buffer retrieval
This commit includes the change required for
retrieving the page after write. The previous
mechanism used was to wait for 10ms post write
and then retrieve it. whereas, now we have used
ptrace's singlestep api to cause a trigger in the
next instruction after write where we shall capture
the written buffer