Test branch for Helloworld enclave
This commit includes the changes required for testing the
helloworld openSGX application with stack migration feature.
Please note that the popSGX-monitor application code commited
here is just a prototype version. Furthermore, also note that
before testing the helloworld application, the ASLR & stack canary
feature in the host and remote node has to be disabled!!!